ohkey

+++ components/Card/Card.tsx
@@ -1,16 +1,25 @@
 'use client'
 import { Button, Card } from "flowbite-react"
-
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faChevronDown, faChevronUp } from '@fortawesome/free-solid-svg-icons';
 import "./styles.css"
+import { useState } from "react";
+import { GameCards } from "./dataCard";
+
 
 export default function ComponentCard() {
-    
+    const [showAllCards, setShowAllCards] = useState(false);
+
     const TypeGame = (Game : any) => {
-        sessionStorage.clear();
-        sessionStorage.setItem('TypeGame' , Game)
+        if (typeof window !== 'undefined') {
+            sessionStorage.clear();
+            sessionStorage.setItem('TypeGame' , Game)
+        }
         
     }
+
     return(
+        
         <div>
             <div className="ml-4 text-orange-400 text-xl font-serif font-bold">Game Populer</div>
             <div className="text-2xl font-extrabold text-white ml-5 mb-10">Trending Sekarang</div>
@@ -75,80 +84,38 @@ export default function ComponentCard() {
 
         <div className="ml-4 text-orange-400 text-xl font-serif mt-10 font-bold">All Games</div>
         <div className="text-2xl font-extrabold text-white ml-5 mb-10">Semua Games Kami</div>
-        <div className="grid grid-cols-3 md:grid-cols-8 gap-4 m-3 btn-card ComCard mt-5">
-        <Card className="h-auto max-w-sm rounded-xl ComCard"
-         imgAlt=""
-         href="/games/mobile-legends"
-         onClick={() => TypeGame("Mobile Legends")}
-         imgSrc="https://cdn1.codashop.com/S/content/mobile/images/product-tiles/ID_MLBB-M4-Codacash-tile.jpg"
-         >
-         <Button className="btn-card font-extrabold ComCard" color="gray">Topup</Button>
-        </Card>
+        <div>
+      <div className="grid grid-cols-3 md:grid-cols-8 gap-4 m-3 btn-card ComCard mt-5">
+        {GameCards.slice(0, showAllCards ? GameCards.length : 3).map((card, index) => (
+          <Card className="h-auto max-w-sm rounded-xl ComCard"
+            key={index}
+            imgAlt={card.imgAlt}
+            href={card.href}
+            imgSrc={card.imgSrc}
+            onClick={() => TypeGame(card.onclick)}
+          >
+            <Button className="btn-card font-extrabold ComCard" color="gray">Topup</Button>
+          </Card>
+        ))}
+      </div>
 
-        <Card className="h-auto max-w-sm rounded-xl ComCard"
-         imgAlt="FreeFire"
-         href="/games/freefire"
-         onClick={() => TypeGame("Free Fire")}
-
-         imgSrc="https://cdn.unipin.com/images/icon_product_pages/1658817763-icon-200x200_icon%20ff.jpg"
-         >
-        <Button className="btn-card font-extrabold ComCard" color="gray">Topup</Button>
-        </Card>
-
-        <Card className="h-auto max-w-sm rounded-xl ComCard"
-         imgAlt=""
-         href="/clashofclans"
-         onClick={() => TypeGame("Clash Of Clans")}
-         imgSrc="https://cdn1.codashop.com/S/content/mobile/images/product-tiles/ID_Clash-of-Clans-20%20(1).jpg"
-         >
-        <Button className="btn-card font-extrabold ComCard" color="gray">Topup</Button>
-        </Card>
-
-        <Card className="h-auto max-w-sm rounded-xl ComCard"
-         imgAlt=""
-         href="/toweroffantasy"
-         onClick={() => TypeGame("Tower Of Fantasy")}
-         imgSrc="https://cdn.unipin.com/images/icon_product_pages/1663645620-icon-1662619195-icon-1662082730-icon-Tower%20of%20Fantasy%20logo%20-%201%20jpg.jpg"
-         >
-        <Button className="btn-card font-extrabold ComCard" color="gray">Topup</Button>
-        </Card>
-
-        <Card className="h-auto max-w-sm rounded-xl ComCard"
-         imgAlt=""
-         href="/CODM"
-         onClick={() => TypeGame("Call Of Duty M")}
-         imgSrc="https://cdn.unipin.com/images/icon_product_pages/1633599388-icon-Icon_1024.jpg"
-         >
-        <Button className="btn-card font-extrabold ComCard" color="gray">Topup</Button>
-        </Card>
-
-        <Card className="h-auto max-w-sm rounded-xl ComCard"
-         imgAlt=""
-         href="/undawn"
-         onClick={() => TypeGame("Undawn")}
-         imgSrc="https://cdn.unipin.com/images/icon_product_pages/1689159121-icon-garena-undawn.jpg"
-         >
-        <Button className="btn-card font-extrabold ComCard" color="gray">Topup</Button>
-        </Card>
-
-        <Card className="h-auto max-w-sm rounded-xl ComCard"
-         imgAlt=""
-         href="/metalslug-awakening"
-         onClick={() => TypeGame("Metal Slug -Awakening")}
-         imgSrc="https://cdn.unipin.com/images/icon_product_pages/1691644925-icon-App%20Icon%20200x200_11zon.jpg"
-         >
-        <Button className="btn-card font-extrabold ComCard" color="gray">Topup</Button>
-        </Card>
-
-        <Card className="h-auto max-w-sm rounded-xl ComCard"
-         imgAlt=""
-         href="/ragnarok-origin"
-         imgSrc="https://cdn.unipin.com/images/icon_product_pages/1680764749-icon-%E5%9B%BE%E7%89%87_20230406150453.jpg"
-         >
-        <Button className="btn-card font-extrabold ComCard" color="gray">Topup</Button>
-        </Card>
-
-        </div>
+      <div className="flex flex-col items-center bg-white rounded-xl mt-5">
+      <Button
+        className="btn-card font-extrabold ComCard"
+        onClick={() => setShowAllCards(!showAllCards)}
+      >
+        {showAllCards ? (
+          <>
+            Lebih Sedikit... <FontAwesomeIcon icon={faChevronUp} />
+          </>
+        ) : (
+          <>
+            Lebih Banyak... <FontAwesomeIcon icon={faChevronDown} />
+          </>
+        )}
+      </Button>
+      </div>
+    </div>
 
           {/* element ke dua */}
           <div className="ml-4 text-orange-400 text-xl font-serif mt-10 font-bold">Your favorit</div>
